+++ crates/wdl-engine/src/backend.rs
@@ -22,6 +22,8 @@ use tokio::sync::oneshot::Receiver;
 use tokio::task::JoinSet;
 use tokio_util::sync::CancellationToken;
 use tracing::debug;
+use wdl_ast::Diagnostic;
+use wdl_ast::v1::TaskDefinition;
 
 use crate::ContentKind;
 use crate::EvaluationPath;
@@ -29,6 +31,7 @@ use crate::GuestPath;
 use crate::Value;
 use crate::http::Location;
 use crate::http::Transferer;
+use crate::tree::SyntaxNode;
 
 mod apptainer;
 mod docker;
@@ -327,9 +330,10 @@ pub(crate) trait TaskExecutionBackend: Send + Sync {
     /// environment or if the task specifies invalid requirements.
     fn constraints(
         &self,
+        task: &TaskDefinition<SyntaxNode>,
         requirements: &HashMap<String, Value>,
         hints: &HashMap<String, Value>,
-    ) -> Result<TaskExecutionConstraints>;
+    ) -> Result<TaskExecutionConstraints, Diagnostic>;
 
     /// Gets the guest (container) inputs directory of the backend.
     ///