+++ crates/wdl-engine/src/backend/docker.rs
@@ -30,6 +30,8 @@ use tokio_util::sync::CancellationToken;
 use tracing::info;
 use tracing::warn;
 use url::Url;
+use wdl_ast::Diagnostic;
+use wdl_ast::v1::TaskDefinition;
 
 use super::TaskExecutionBackend;
 use super::TaskExecutionConstraints;
@@ -50,12 +52,17 @@ use crate::config::Config;
 use crate::config::DEFAULT_TASK_SHELL;
 use crate::config::DockerBackendConfig;
 use crate::config::TaskResourceLimitBehavior;
+use crate::tree::SyntaxNode;
 use crate::v1::container;
 use crate::v1::cpu;
+use crate::v1::cpu_from_values;
 use crate::v1::gpu;
 use crate::v1::max_cpu;
+use crate::v1::max_cpu_from_values;
 use crate::v1::max_memory;
+use crate::v1::max_memory_from_values;
 use crate::v1::memory;
+use crate::v1::memory_from_values;
 
 /// The root guest path for inputs.
 const GUEST_INPUTS_DIR: &str = "/mnt/task/inputs/";
@@ -343,13 +350,15 @@ impl TaskExecutionBackend for DockerBackend {
 
     fn constraints(
         &self,
+        task: &TaskDefinition<SyntaxNode>,
         requirements: &HashMap<String, Value>,
         hints: &HashMap<String, Value>,
-    ) -> Result<TaskExecutionConstraints> {
+    ) -> Result<TaskExecutionConstraints, Diagnostic> {
         let container = container(requirements, self.config.task.container.as_deref());
 
-        let mut cpu = cpu(requirements);
-        if (self.max_cpu as f64) < cpu {
+        let mut required_cpu = cpu(task, requirements);
+        if (self.max_cpu as f64) < required_cpu.value {
+            let span = required_cpu.span;
             let env_specific = if self.config.suppress_env_specific_output {
                 String::new()
             } else {
@@ -362,22 +371,27 @@ impl TaskExecutionBackend for DockerBackend {
                 TaskResourceLimitBehavior::TryWithMax => {
                     warn!(
                         "task requires at least {cpu} CPU{s}{env_specific}",
-                        s = if cpu == 1.0 { "" } else { "s" },
+                        cpu = required_cpu.value,
+                        s = if required_cpu.value == 1.0 { "" } else { "s" },
                     );
                     // clamp the reported constraint to what's available
-                    cpu = self.max_cpu as f64;
+                    required_cpu.value = self.max_cpu as f64;
                 }
                 TaskResourceLimitBehavior::Deny => {
-                    bail!(
+                    let msg = format!(
                         "task requires at least {cpu} CPU{s}{env_specific}",
-                        s = if cpu == 1.0 { "" } else { "s" },
+                        cpu = required_cpu.value,
+                        s = if required_cpu.value == 1.0 { "" } else { "s" },
                     );
+                    return Err(Diagnostic::error(msg)
+                        .with_label("this requirement exceeds the available CPUs", span));
                 }
             }
         }
 
-        let mut memory = memory(requirements)?;
-        if self.max_memory < memory as u64 {
+        let mut required_memory = memory(task, requirements)?;
+        if self.max_memory < required_memory.value as u64 {
+            let span = required_memory.span;
             let env_specific = if self.config.suppress_env_specific_output {
                 String::new()
             } else {
@@ -391,17 +405,91 @@ impl TaskExecutionBackend for DockerBackend {
                     warn!(
                         "task requires at least {memory} GiB of memory{env_specific}",
                         // Display the error in GiB, as it is the most common unit for memory
-                        memory = memory as f64 / ONE_GIBIBYTE,
+                        memory = required_memory.value as f64 / ONE_GIBIBYTE,
                     );
                     // clamp the reported constraint to what's available
-                    memory = self.max_memory.try_into().unwrap_or(i64::MAX);
+                    required_memory.value = self.max_memory.try_into().unwrap_or(i64::MAX);
                 }
                 TaskResourceLimitBehavior::Deny => {
-                    bail!(
+                    let msg = format!(
                         "task requires at least {memory} GiB of memory{env_specific}",
                         // Display the error in GiB, as it is the most common unit for memory
-                        memory = memory as f64 / ONE_GIBIBYTE,
+                        memory = required_memory.value as f64 / ONE_GIBIBYTE,
                     );
+                    return Err(Diagnostic::error(msg)
+                        .with_label("this requirement exceeds the available memory", span));
+                }
+            }
+        }
+
+        if let Some(required_max_cpu) = max_cpu(task, hints)
+            && (self.max_cpu as f64) < required_max_cpu.value
+        {
+            let span = required_max_cpu.span;
+            let env_specific = if self.config.suppress_env_specific_output {
+                String::new()
+            } else {
+                format!(
+                    ", but the execution backend has a maximum of {max}",
+                    max = self.max_cpu
+                )
+            };
+            match self.config.task.cpu_limit_behavior {
+                TaskResourceLimitBehavior::TryWithMax => {
+                    warn!(
+                        "task requests a maximum of {mcpu} CPU{s}{env_specific}",
+                        mcpu = required_max_cpu.value,
+                        s = if required_max_cpu.value == 1.0 {
+                            ""
+                        } else {
+                            "s"
+                        }
+                    );
+                }
+                TaskResourceLimitBehavior::Deny => {
+                    let msg = format!(
+                        "task requests a maximum of {mcpu} CPU{s}{env_specific}",
+                        mcpu = required_max_cpu.value,
+                        s = if required_max_cpu.value == 1.0 {
+                            ""
+                        } else {
+                            "s"
+                        }
+                    );
+                    return Err(
+                        Diagnostic::error(msg).with_label("this hint exceeds available CPUs", span)
+                    );
+                }
+            }
+        }
+
+        let required_max_memory = max_memory(task, hints)?;
+        if let Some(mmem) = required_max_memory.as_ref()
+            && self.max_memory < mmem.value as u64
+        {
+            let span = mmem.span;
+            let env_specific = if self.config.suppress_env_specific_output {
+                String::new()
+            } else {
+                format!(
+                    ", but the execution backend has a maximum of {max_memory} GiB",
+                    max_memory = self.max_memory as f64 / ONE_GIBIBYTE
+                )
+            };
+            match self.config.task.cpu_limit_behavior {
+                TaskResourceLimitBehavior::TryWithMax => {
+                    warn!(
+                        "task requests a maximum of {memory} GiB of memory{env_specific}",
+                        memory = mmem.value as f64 / ONE_GIBIBYTE
+                    );
+                }
+                TaskResourceLimitBehavior::Deny => {
+                    let msg = format!(
+                        "task requests a maximum of {memory} GiB of memory{env_specific}",
+                        memory = mmem.value as f64 / ONE_GIBIBYTE,
+                    );
+                    return Err(Diagnostic::error(msg)
+                        .with_label("this hint exceeds the available memory", span));
                 }
             }
         }
@@ -417,8 +505,8 @@ impl TaskExecutionBackend for DockerBackend {
 
         Ok(TaskExecutionConstraints {
             container: Some(container.into_owned()),
-            cpu,
-            memory,
+            cpu: required_cpu.value,
+            memory: required_memory.value,
             gpu,
             fpga: Default::default(),
             disks: Default::default(),
@@ -444,16 +532,22 @@ impl TaskExecutionBackend for DockerBackend {
         let hints = request.hints();
 
         let container = container(requirements, self.config.task.container.as_deref()).into_owned();
-        let mut cpu = cpu(requirements);
+        let mut cpu = cpu_from_values(requirements);
         if let TaskResourceLimitBehavior::TryWithMax = self.config.task.cpu_limit_behavior {
             cpu = std::cmp::min(cpu.ceil() as u64, self.max_cpu) as f64;
         }
-        let mut memory = memory(requirements)? as u64;
+        let mut memory = memory_from_values(requirements)? as u64;
         if let TaskResourceLimitBehavior::TryWithMax = self.config.task.memory_limit_behavior {
             memory = std::cmp::min(memory, self.max_memory);
         }
-        let max_cpu = max_cpu(hints);
-        let max_memory = max_memory(hints)?.map(|i| i as u64);
+        let mut max_cpu = max_cpu_from_values(hints);
+        if let TaskResourceLimitBehavior::TryWithMax = self.config.task.cpu_limit_behavior {
+            max_cpu = max_cpu.map(|mcpu| f64::min(mcpu, self.max_cpu as f64));
+        }
+        let mut max_memory = max_memory_from_values(hints)?.map(|i| i as u64);
+        if let TaskResourceLimitBehavior::TryWithMax = self.config.task.memory_limit_behavior {
+            max_memory = max_memory.map(|mmem| std::cmp::min(mmem, self.max_memory));
+        }
         let gpu = gpu(requirements, hints);
 
         let name = format!(