+++ crates/wdl-engine/src/backend/slurm_apptainer.rs
@@ -31,6 +31,7 @@ use tokio_util::sync::CancellationToken;
 use tracing::debug;
 use tracing::trace;
 use tracing::warn;
+use wdl_ast::Diagnostic;
 
 use super::ApptainerState;
 use super::TaskExecutionBackend;
@@ -45,6 +46,7 @@ use crate::backend::TaskExecutionResult;
 use crate::config::Config;
 use crate::config::SlurmApptainerBackendConfig;
 use crate::config::TaskResourceLimitBehavior;
+use crate::tree::SyntaxNode;
 use crate::v1;
 
 /// The name of the file where the Apptainer command invocation will be written.
@@ -402,11 +404,16 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
 
     fn constraints(
         &self,
+        task: &wdl_ast::v1::TaskDefinition<SyntaxNode>,
         requirements: &HashMap<String, Value>,
         hints: &HashMap<String, crate::Value>,
-    ) -> anyhow::Result<super::TaskExecutionConstraints> {
-        let mut required_cpu = v1::cpu(requirements);
-        let mut required_memory = ByteSize::b(v1::memory(requirements)? as u64);
+    ) -> anyhow::Result<super::TaskExecutionConstraints, Diagnostic> {
+        let mut required_cpu = v1::cpu(task, requirements);
+        let required_memory = v1::memory(task, requirements)?;
+        let (mut required_memory, required_memory_span) = (
+            ByteSize::b(required_memory.value as u64),
+            required_memory.span,
+        );
 
         // Determine whether CPU or memory limits are set for this partition, and clamp
         // or deny them as appropriate if the limits are exceeded
@@ -419,8 +426,9 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
             .slurm_partition_for_task(requirements, hints)
         {
             if let Some(max_cpu) = partition.max_cpu_per_task()
-                && required_cpu > max_cpu as f64
+                && required_cpu.value > max_cpu as f64
             {
+                let span = required_cpu.span;
                 let env_specific = if self.engine_config.suppress_env_specific_output {
                     String::new()
                 } else {
@@ -430,22 +438,27 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
                     TaskResourceLimitBehavior::TryWithMax => {
                         warn!(
                             "task requires at least {required_cpu} CPU{s}{env_specific}",
-                            s = if required_cpu == 1.0 { "" } else { "s" },
+                            required_cpu = required_cpu.value,
+                            s = if required_cpu.value == 1.0 { "" } else { "s" },
                         );
                         // clamp the reported constraint to what's available
-                        required_cpu = max_cpu as f64;
+                        required_cpu.value = max_cpu as f64;
                     }
                     TaskResourceLimitBehavior::Deny => {
-                        bail!(
+                        let msg = format!(
                             "task requires at least {required_cpu} CPU{s}{env_specific}",
-                            s = if required_cpu == 1.0 { "" } else { "s" },
+                            required_cpu = required_cpu.value,
+                            s = if required_cpu.value == 1.0 { "" } else { "s" },
                         );
+                        return Err(Diagnostic::error(msg)
+                            .with_label("this requirement exceeds the available CPUs", span));
                     }
                 }
             }
             if let Some(max_memory) = partition.max_memory_per_task()
                 && required_memory > max_memory
             {
+                let span = required_memory_span;
                 let env_specific = if self.engine_config.suppress_env_specific_output {
                     String::new()
                 } else {
@@ -464,10 +477,12 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
                         required_memory = max_memory;
                     }
                     TaskResourceLimitBehavior::Deny => {
-                        bail!(
+                        let msg = format!(
                             "task requires at least {required_memory} GiB of memory{env_specific}",
                             required_memory = required_memory.as_u64() as f64 / ONE_GIBIBYTE
                         );
+                        return Err(Diagnostic::error(msg)
+                            .with_label("this requirement exceeds the available memory", span));
                     }
                 }
             }
@@ -482,7 +497,7 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
             //
             // sinfo -p <partition> -s --json | jq .sinfo[0].cpus
             // sinfo -p <partition> -s --json | jq .sinfo[0].memory
-            cpu: required_cpu,
+            cpu: required_cpu.value,
             memory: required_memory.as_u64().try_into().unwrap_or(i64::MAX),
             // TODO ACF 2025-10-16: these are almost certainly wrong
             gpu: Default::default(),
@@ -512,8 +527,8 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
         let container =
             v1::container(requirements, self.engine_config.task.container.as_deref()).into_owned();
 
-        let mut required_cpu = v1::cpu(requirements);
-        let mut required_memory = ByteSize::b(v1::memory(requirements)? as u64);
+        let mut required_cpu = v1::cpu_from_values(requirements);
+        let mut required_memory = ByteSize::b(v1::memory_from_values(requirements)? as u64);
 
         // Determine whether CPU or memory limits are set for this partition, and clamp
         // or deny them as appropriate if the limits are exceeded
@@ -582,8 +597,8 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
 
         // TODO ACF 2025-10-23: investigate whether Slurm offers hard vs soft limits for
         // CPU and memory
-        let _max_cpu = v1::max_cpu(hints);
-        let _max_memory = v1::max_memory(hints)?.map(|i| i as u64);
+        let _max_cpu = v1::max_cpu_from_values(hints);
+        let _max_memory = v1::max_memory_from_values(hints)?.map(|i| i as u64);
 
         // Truncate the request ID to fit in the Slurm job name length limit.
         let request_id = request.id();